# docs/swagger.yaml
definitions:
...
  mengri-flow_internal_app_dto.AddWorkspaceMemberRequest:
    properties:
      accountId:
        type: string
      role:
        enum:
        - member
        - admin
        type: string
    required:
    - accountId
    - role
    type: object
...
  mengri-flow_internal_app_dto.CreateTriggerRequest:
    properties:
      clusterId:
        type: string
      config:
        additionalProperties: true
        type: object
      description:
        maxLength: 500
        type: string
      errorHandling:
        $ref: '#/definitions/mengri-flow_internal_app_dto.ErrorHandlingConfig'
      flowId:
        type: string
      flowVersion:
        type: integer
      inputMapping:
        additionalProperties:
          type: string
        type: object
      name:
        maxLength: 100
        type: string
      outputMapping:
        additionalProperties:
          type: string
        type: object
      type:
        type: string
      workspaceId:
        type: string
    required:
    - config
    - flowId
    - name
    - type
    type: object
  mengri-flow_internal_app_dto.CreateWorkspaceRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
  mengri-flow_internal_app_dto.ErrorHandlingConfig:
    properties:
      customErrorFormat:
        additionalProperties: true
        type: object
      retryOnFailure:
        type: boolean
      strategy:
        type: string
    type: object
  mengri-flow_internal_app_dto.ErrorHandlingResponse:
    properties:
      customErrorFormat:
        additionalProperties: true
        type: object
      retryOnFailure:
        type: boolean
      strategy:
        type: string
    type: object
...
  mengri-flow_internal_app_dto.ListTriggersResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/mengri-flow_internal_app_dto.TriggerResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  mengri-flow_internal_app_dto.ListWorkspacesResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/mengri-flow_internal_app_dto.WorkspaceResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  mengri-flow_internal_app_dto.PublishTriggerRequest:
    properties:
      clusterId:
        type: string
      triggerId:
        type: string
    required:
    - clusterId
    - triggerId
    type: object
...
  mengri-flow_internal_app_dto.TriggerResponse:
    properties:
      clusterId:
        type: string
      config:
        additionalProperties: true
        type: object
      createdAt:
        type: string
      description:
        type: string
      errorHandling:
        $ref: '#/definitions/mengri-flow_internal_app_dto.ErrorHandlingResponse'
      flowId:
        type: string
      flowVersion:
        type: integer
      id:
        type: string
      inputMapping:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      outputMapping:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
...
  mengri-flow_internal_app_dto.UpdateTriggerRequest:
    properties:
      clusterId:
        type: string
      config:
        additionalProperties: true
        type: object
      description:
        maxLength: 500
        type: string
      errorHandling:
        $ref: '#/definitions/mengri-flow_internal_app_dto.ErrorHandlingConfig'
      flowVersion:
        type: integer
      inputMapping:
        additionalProperties:
          type: string
        type: object
      name:
        maxLength: 100
        type: string
      outputMapping:
        additionalProperties:
          type: string
        type: object
    type: object
  mengri-flow_internal_app_dto.UpdateWorkspaceRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
    type: object
  mengri-flow_internal_app_dto.WorkspaceMemberResponse:
    properties:
      accountId:
        type: string
      displayName:
        type: string
      email:
        type: string
      joinedAt:
        type: string
      role:
        type: string
    type: object
  mengri-flow_internal_app_dto.WorkspaceResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      memberCount:
        type: integer
      name:
        type: string
      ownerId:
        type: string
      updatedAt:
        type: string
    type: object
  mengri-flow_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /triggers:
    get:
      consumes:
      - application/json
      parameters:
      - description: 流程ID
        in: query
        name: flowId
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ListTriggersResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取触发器列表
      tags:
      - Trigger
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建触发器请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.CreateTriggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.TriggerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建触发器
      tags:
      - Trigger
  /triggers/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 触发器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除触发器
      tags:
      - Trigger
    get:
      consumes:
      - application/json
      parameters:
      - description: 触发器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.TriggerResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取触发器详情
      tags:
      - Trigger
    put:
      consumes:
      - application/json
      parameters:
      - description: 触发器ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新触发器请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.UpdateTriggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.TriggerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新触发器
      tags:
      - Trigger
  /triggers/{id}/disable:
    post:
      consumes:
      - application/json
      parameters:
      - description: 触发器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 禁用触发器
      tags:
      - Trigger
  /triggers/{id}/enable:
    post:
      consumes:
      - application/json
      parameters:
      - description: 触发器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 启用触发器
      tags:
      - Trigger
  /triggers/{id}/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: 触发器ID
        in: path
        name: id
        required: true
        type: string
      - description: 发布请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.PublishTriggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 发布触发器到集群
      tags:
      - Trigger
  /workspaces:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ListWorkspacesResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取工作空间列表
      tags:
      - Workspace
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建工作空间请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.CreateWorkspaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.WorkspaceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建工作空间
      tags:
      - Workspace
  /workspaces/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 工作空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除工作空间
      tags:
      - Workspace
    get:
      consumes:
      - application/json
      parameters:
      - description: 工作空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.WorkspaceResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取工作空间详情
      tags:
      - Workspace
    put:
      consumes:
      - application/json
      parameters:
      - description: 工作空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新工作空间请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.UpdateWorkspaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.WorkspaceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新工作空间
      tags:
      - Workspace
  /workspaces/{id}/members:
    post:
      consumes:
      - application/json
      parameters:
      - description: 工作空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 添加成员请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.AddWorkspaceMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.WorkspaceMemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 添加工作空间成员
      tags:
      - Workspace
  /workspaces/{id}/members/{userId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 工作空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 成员ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 移除工作空间成员
      tags:
      - Workspace
swagger: "2.0"